@@ -9,22 +9,22 @@
     "copy-shiki-assets": "cp -R ../../node_modules/shiki/{themes,languages} ./public/assets/shiki"
   },
   "dependencies": {
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "prettier": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "prettier": "^3.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-figma-plugin-ds": "^2.3.0",
     "shiki": "^0.9.3"
   },
   "devDependencies": {
-    "@types/css-font-loading-module": "^0.0.4",
-    "@types/prettier": "^2.3.0",
-    "@vitejs/plugin-react-refresh": "^1.3.3",
-    "autoprefixer": "^10.2.6",
-    "postcss": "^8.3.3",
-    "tailwindcss": "^2.1.4",
-    "typescript": "^4.3.2",
-    "vite": "^2.3.7"
+    "@types/css-font-loading-module": "^0.0.9",
+    "@types/prettier": "^3.0.0",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "autoprefixer": "^10.4.15",
+    "postcss": "^8.4.29",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9"
   }
 }
