@@ -1,20 +1,19 @@
 {
   "version": "0.0.1",
   "dependencies": {
-    "@figma/plugin-typings": "^1.24.0",
+    "@figma/plugin-typings": "^1.76.0",
     "@internal/plugin-shared": "workspace:*",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "monaco-editor": "0.25.0",
-    "node-fetch": "^2.6.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "estrella": "^1.4.1",
-    "typescript": "^4.3.2"
+    "typescript": "^5.2.2"
   },
   "scripts": {
     "dev": "tsc -b && node build.js -w",
