@@ -1,8 +1,9 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import Main from './components/Main'
-
 import './index.css'
 import 'react-figma-plugin-ds/figma-plugin-ds.css'
 
-ReactDOM.render(<Main />, document.getElementById('root'))
+import React from 'react'
+import { createRoot } from 'react-dom/client'
+
+import Main from './components/Main'
+
+createRoot(document.getElementById('root')!).render(<Main />)
