@@ -76,6 +76,8 @@ const Main: React.FC = () => {
         availableFigmaFonts.resolve(event.data.pluginMessage.monoFontFamilies)
       }
     }
+
+    parent.postMessage({ pluginMessage: { type: 'INIT' } }, '*')
   }, [setCode, setFontSize, availableFigmaFonts])
 
   const execRun = async () => {
@@ -95,8 +97,10 @@ const Main: React.FC = () => {
   }
 
   const runPrettier = async () => {
+    // NOTE we're lazy-loading prettier in order to speed up the initial load time
     const { formatText } = await import('../utils/prettier')
-    setCode(formatText(code, language))
+    const formattedCode = await formatText(code, language)
+    setCode(formattedCode)
   }
 
   if (fontRes.isLoading || isHighlighterLoading) {
@@ -126,6 +130,7 @@ const Main: React.FC = () => {
           code: code,
           setCode: setCode,
           lineHeight: fontSize * 1.5,
+          env,
         }}
       />
       <Sidebar
