@@ -1,13 +1,14 @@
 import type { BuiltInParserName } from 'prettier'
-import babel from 'prettier/parser-babel'
+import babel from 'prettier/plugins/babel'
+import estree from 'prettier/plugins/estree'
 import prettier from 'prettier/standalone'
 import type { Lang } from 'shiki'
 
-export const formatText = (text: string, language: Lang): string => {
+export const formatText = (text: string, language: Lang): Promise<string> => {
   return prettier.format(text, {
     // TODO needs to be more selective whether prettier supports a given language
     parser: getParser(language),
-    plugins: [babel],
+    plugins: [estree, babel],
     singleQuote: true,
   })
 }
